Extract APIEndpoint protocols
